extract parallelize search server per region
